harden file upload and check against magic bytes instead of trusting file extensions, add no-sniff header

+++ lib/mindwendel_web/live/lane_live/index_component.ex
@@ -12,7 +12,8 @@ defmodule MindwendelWeb.LaneLive.IndexComponent do
 
     %{current_user: current_user, brainstorming: brainstorming} = socket.assigns
 
-    if lane && has_moderating_permission(brainstorming.id, current_user) do
+    if lane && lane.brainstorming_id == brainstorming.id &&
+         has_moderating_permission(brainstorming.id, current_user) do
       {:ok, _} = Lanes.delete_lane(lane)
     end
